Move config repository and redactor to infrastructure
